Increase the audit interval of device image update
This method verifies the timeout of the update process of the images,
and the minimum timeout available inside the code is defined in the
fw_update_small_timeout variable. This variable is not changed by any
other process and its value is set to 300 seconds.

After a research, we noticed that the appropriate time interval
for this audit is to be executed between 300 seconds interval.

This commit is a continuation of the study started at task 45616.

Note: This value is set by Puppet (story 45777). The only reason that
it's being updated here is to keep the same value of Puppet.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45776

Depends-On: https://review.opendev.org/c/starlingx/config/+/848437
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: Id3d8cdf78dafb1e15c542159ed1288fa5d63a7a0