Increase the audit interval of storage backend failure
This audit verifies if there's a storage backend stuck in 'configuring'
state, the timeout for this is 1200 seconds, set in the
SB_CONFIGURATION_TIMEOUT constant.

After a research, we noticed that the time interval does not need to be
60 seconds, since the timeout of the configuration is 1200. Increasing
it to 400 seconds (1/3 of the timeout) is enough to capture any storage
backend stucked.

This commit is a continuation of the study started at task 45616.

Note: This value is set by Puppet (story 45777). The only reason that
it's being updated here is to keep the same value of Puppet.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45827

Depends-On: https://review.opendev.org/c/starlingx/config/+/848437
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I8004bb856755b398ebcf52e0640362fdbfc217ed