Refactor of _audit_deferred_runtime_config
The purpose of this audit is to gather several related updates and
execute as a single runtime Puppet manifest apply since the overhead
of individual updates are high.

As this audit only executes when the list
'_host_deferred_runtime_config' is not empty, this commit changes the
verification of the state of this list to the beginning of the code.
With this change, the audit does not need to execute other methods if
the list is empty.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being executed right with the
list '_host_deferred_runtime_config' empty or not.
and that the
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45885

Depends-On: https://review.opendev.org/c/starlingx/config/+/848437
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I628277418717ad3425ce1a3cce127c98df936d22