System CLI interface naming rule inconsistency
Starling-X API does not allow the use of dashes and dots
in the interface name. The aim is to allow the use of dots,
dashes, and underscores in general.

This change is because of Starling-X and K8s having
naming rule inconsistencies and might cause confusion to customer.

Test Plan:

PASS: Obtain the list of  characters that are deemed not valid
      for interface names and remove dots, dashes, and underscores
      from the list.

      Test 1: using "+": system host-if-add ... controller-0 vf+2 vf ..
      --> Cannot use '+' as a special character in interface name.
      Test 2: using "=": system host-if-add ... controller-0 vf=2 vf ..
      --> Cannot use '=' as a special character in interface name.
      Test 3: using ".": system host-if-add ... controller-0 vf.2 vf ..
      --> No error returned
      Test 4: using "-": system host-if-add ... controller-0 vf-2 vf ..
      --> No error returned
      Test 5: using "_": system host-if-add ... controller-0 vf_2 vf ..
      --> No error returned

PASS: The changes were tested on a controller to make sure it can
      be locked and unlocked.

Closes-Bug: 1983614

Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>
Change-Id: Ie50981dfca988bcb47eaf1a28603ad24f2de83d7