Fix variable initialisation error in _audit_install_states
Task 45646 changes the install_state variable name but it was not
necessary.

So this commit fixes it to the correct name.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that the mentioned audit is being called.

Story: 2010087
Task: 45953

Depends-On: https://review.opendev.org/c/starlingx/config/+/848437
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: Ia1483cf4bf3aa70853f3ab9a27aa0220825a54ba