Change periodic_tasks timer to be dynamic instead of fixed
Today the periodic tasks are executed in a fixed time slot of 60
seconds. So, even if we spread the periodic tasks over the time, they
will execute at the same time frequently. This happens because of the
implementation inside the Service module, where there is a method from
ThreadGroup called add_timer.

The purpose of this commit is to change this method, adding the method
add_dynamic_timer from the same module to be executed instead, if
selected. With this change, if the parameter 'periodic_interval' is
set to 0, there is no fixed timeslot and all the periodic tasks will
be executed at the interval that they are configured. Note that if
this parameter is set to a value higher than 0, the execution will
be fixed (as it is today).

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45983

Depends-On: https://review.opendev.org/c/starlingx/config/+/851468
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: Ieffa4c6144228355cf530626ba6ee440a06ac0fb