Added postgres bin path to cmd calls
Postgres bin path is different between centos and debian.

PG has an exec called pg_wrapper which redirects calls
for some of PG's executables, but not for all of them,
causing some command calls to not find the target binary.

Changed command calls to use the fullpath to the binaries
so as to make them more reliable and independent of changes
to the PATH env variable.

Test Plan:
pass: code snippets work on python console on centos and debian
pass: pg cmds find the binary files on both centos and debian
pass: run upgrade on centos with AIO-SX/DX

Story: 2009303
Task: 45988

Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>
Change-Id: I2e0cdaf86dd5665dd1b8cb7558c287016cf249a3