Increase the audit interval of upgrade status to 180 seconds
The audit_upgrade_status method verifies if there is a system upgrade
running. If so, it takes actions based on its state. Note that the
system stays in each state for a long time but the periodicity of this
audit is 60 seconds, causing unecessary executions during each state.

After a research, we noticed that the appropriate time interval
for this audit is to be executed between 180 seconds interval. This
interval was choosen because upgrades are not frequenty done on the
system and they already take time to be executed.

This commit is a continuation of the study started at task 45616.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.
PASS: AIO-SX: perform a system upgrade with this new interval.
PASS: AIO-DX: failed to perform a system upgrade with this new interval
(was expected to fail).

Story: 2010087
Task: 45999

Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I9a0be29401248a0018f9ec89e5e0db7cbfa00c6a