Do not delete secrets on cert manager upgrade
Cert manager is configured to delete secrets when the corresponding
certificate is deleted. As part of data migration during an upgrade,
we delete and recreate cert manager certificates in the new fluxcd
version of cert manager. This causes all certificates to be renewed
due to missing secrets when cert manager is upgraded. This commit
adds a system override to NOT delete secrets when a certificate is
deleted. It also removes the upstream documentation step of deleting
CRDs, since the CRDs are already handled by helm.

Test Cases:

PASS: Upgrade on a setup that pretty reliably reproduces the issue.
      The cert manager certificate secrets were not recreated.
PASS: Manually execute the steps in the script slowly. Without
      the change, secrets were recreated. With the change, they
      were not.

Change-Id: I55715fc4e6f4d2eb9c2e6429801c2b2a2044e122
Closes-Bug: 1985584
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>