Remap partition layout for upgrading to Debian
This is to remap the new partition in N+1 system.

For upgrading to Debian from CentOS, the partition layout changes a lot.
This migration script is to pick up the new layout and create
corresponding sysinv records so that ensure the puppet hieradata can be
created correctly before sysinv-agent report inventory.

Story: 2009303
Task: 46009

Test plan:
    1. passed upgrade controller-1 from CentOS 22.06 to Debian 22.12.
       Controller-1 unlocked successfully.

Change-Id: I2243a37b496b2a8a41c7c8d32ef1aad38dd29001
Signed-off-by: Bin Qian <bin.qian@windriver.com>