Disable nohz_full in a virtual env
In VirtualBox, after unlock, SM has all of its services
in 'initial' state.

The reason for this is that SM will not proceed unless
it detects there are no timer delays.

This is particularly noticable for AIO-SX.

By disabling nohz_full in virtual box, the timers are
not delayed and SM is able to start up its services
more quickly (5 seconds).  Othwerwise SM initialization
on a 4 core system can range from 10 minutes to 10 hours.

Test Plan:
  Build/Bootstrap/Unlock Debian AIO-SX on virtualbox.

Closes-Bug: 1890323
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I94226721d2ccd83a8b0caac09d1c745d4c908ae4