Add keystone database migration script
This commit adds an upgrade script to perform the keystone data
migration. This the migration command will run only if the distro
is debian.

Test Plan:

PASS: Verify the DX upgrade from stx 22.06 to 22.12
PASS: Verify that the script for keystone data-migration runs without
      error
PASS: Verify with a database dump that data is migrated successfully and
      only DDL changes are present.
PASS: Verify that keystone services are up and openstack keystone
      commands are working, such as openstack user list, etc.
PASS: Verify that it's possible to change a user's password using
      keystone and verify that it's possible to log back with that user

Story: 2009303
Task: 46075
Depends-On: https://review.opendev.org/c/starlingx/integ/+/854246

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I9e9496ea6ce19adaac60105f637a29c278951314