debian_iso_image.inc: align the package name with "stx-std.lst"
Currently we use the "stx-std.lst" file in tools to add install package.
We are going to use the "debian_iso_image.inc" file to follow CentOS.

We observed some layers already have the debian_iso_image.inc, so first
of all, we need to check if all the packages are available, otherwise,
once supporting "debian_iso_image.inc", "build-image" will fail.

Comment "storageconfig" is to align with "stx-std.lst" in which the
package is commented due to integration issue.

In debian build, only one binary "workerconfig-standalone" is built from
source, so change "workerconfig" to "workerconfig-standalone" which also
align with "build-image"

Story: 2008846
Task: 46141

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: Ife3cecbced143f9f0172198ece2a8f9aee6ed058