Increase timed out for cert-mon certificate installation
Cert-mon can fail to install certificates during host-unlock because
this can be a busy moment for the system, especially for sysinv. This
happens when the system is under stress with high CPU load ( usually
in virtual environments where many vms are running in parallel).

This change only affects the cert-mon rest client.

Test plan:

PASS:  Check if cert-mon is able to install certificates on host-unlock
       under high CPU load

Story: 2009834
Task: 46166
Change-Id: I4ae92aa8d07e717401e815ac156dc60233383e14
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>