Expose the Hyper-Threading (HT) processor feature
This commit implements a method to verify if the Hyper-Threading (HT)
feature is enabled on host. This method will be used in Task 45621 to
compute the eng_workers parameter.

TEST PLAN:
PASS: AIO-SX: create and deploy CentOS installation ISO successfully
on host with HT enabled.
PASS: AIO-SX: check if the correct qty of threads (in this case only
one) was configured on base services.

PASS: AIO-SX: create and deploy CentOS installation ISO successfully
on host with HT disabled.
PASS: AIO-SX: check if the correct qty of threads (2) was configured
on base services.

PASS: AIO-SX: change the number of dedicated platform cores using the
CLI and verify the change was effective and reflected in the services.

Story: 2010087
Task: 46179

Signed-off-by: Reynaldo P Gomes <reynaldo.patronegomes@windriver.com>
Change-Id: I95cf3d21316307001f42053ba0105c2bd3416fcc