Debian: Update the device image cache directory
The current device image cache directory is off /usr which is
a read-only filesystem in Debian. It is moved to /var, a read-write
filesystem, so that image cache directory can be created during
firmware update on Debian. The directory structure is unchanged for
CentOS.

Test Plan:
Verified: The device image cache directory can be created during
          FW update on Debian. And the code changes would not
          impact CentOS.

Story: 2010119
Task: 46230

Signed-off-by: lzhu1 <li.zhu@windriver.com>
Change-Id: Ic1c4db352a0032395dab853cbfb7f886e009cbfa