Create openldap certificate on upgrade
This commit:

Adds an upgrade script to call the ansible playbook
upgrade-openldap-certificate.yml and create the openldap certificate
during the upgrade activate stage.

Changes the puppet plugin code for ldap to read the openldap
certificate only in activate statuses during an upgrade.

Fixes an issue with certificate_shell where it may call len() for a
certificate that does not have the subject information, resulting a
NoneType error.

Test Plan:

PASS: Upgrade from stx 22.06 to 22.12 and verify that the openldap
      certificate is successfully created in kubernetes
      in upgrade activate stage
PASS: After certificate is created in kubernetes, verify that the
      certificate and key is read by puppet and saved to files
      in /etc/ldap/certs
PASS: After certificate is created in kubernetes, verify that the
      certificate is picked up by cert-mon and saved to database
      and show in 'system certificate-list'

Story: 2009834
Task: 45933

Change-Id: Ia4b9d1921b1e7afdc29f398f902faf2a8bf1e25b
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/856556