Fix certificate ssl_ca cert install by dc-orch sync
This commit fixes an issue where trying to install the same certificate
again results in a 'Cannot install certificate with same subject'. That
is incorrect and should be thrown only for a different certificate with
the same subject.

Test Plan:

PASS: Manage a subcloud and verify that it's able to synchronize certs
      without the 'Cannot install certificate with same subject' error
PASS: Try to install the same certificate multiple times and verify
      that no 'Cannot install certificate with same subject' error
      is returned
PASS: Try to install two different certificates with same subjects and
      verify that a 'Cannot install certificate with same subject' error
      is returned

Closes-Bug: 1990007

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I17861145f20b8e1ef61896c3271a96a28fe9ded2