Better error message for sysinv forbidden error
This commit catches the forbidden error raised for a user trying to
run sysinv commands without enough privileges. The forbidden exception
used to not be caught, resulting in sysinv CLI returning a "None" to
the user. With this commit, a more useful error message is shown to
the user.

Test Cases:

PASS: Create a reader user and run "system modify --description".
      Ensure a meaningful error is returned since readers are not
      allowed to run system modify. Ensure no changes are made
      by checking "system show"

Change-Id: I11c407f06196962ba445c6d8a9f7591cc8a5cf05
Story: 2010149
Task: 46360
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>