Replace unsafe Popen when executing playbooks
If the sysinv process is killed while a playbook is executing
in a subprocess that was created with this Popen, it will become
a zombie.

By replacing Popen with run, the aforementioned risk is mitigated
because of the builtin context management and process reaping.

TEST PLAN
PASS: Reboot a deployed system (SX/DX), normal operations expected

Closes-Bug: 1990547
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I1c73ae01e4fa0d61fe774607b7438480470b096c