create vim_db locally and move to nfs device
This change is to avoid intermittent file lock error when
creating vim database directly on nfs device.

As a safer (and more efficient) way is to create the database
on local temp directory and copy it to the nfs mount path.

Also add an audit code to determine if the database copied to
nfs still has the file lock issue and report in log.

Note that the database does not need to be opened over nfs mount,
so the file lock failure issue would not impact the system.

TCs:
   passed DX upgrade 22.06 to 22.12 Debian completed.

Closes-Bug: 1990544
Change-Id: Ib3f1dee3df4f0c240c919b3f5c3414a6b807b1de
Signed-off-by: Bin Qian <bin.qian@windriver.com>