Fixed AttributeError handling in rpc dispatcher
Exception handling is failing to catch an AttributeError
exception. Code snippet is replaced here by an equivalent
approach using a built-in python function.

The message always comes coupled with another error message.
It is an artifact of the appfwk and provides
no information about the root cause or how to debug it.
Should be removed to avoid confusion.

Test Plan:
PASS Following steps to reproduce the message no longer
     trigger it.

Closes-Bug: 1991701
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>
Change-Id: Ia37cf4c5340b58e7f318683669eaf35871558134