Debian: fix oam VLAN interface MTU
When Debian's ifup tool runs for a IPv6 VLAN interface it is not setting
the MTU found in the configuration file. Instead it sets it to the
underlying interface's MTU. If that's a jumbo MTU value, it can cause
packet drops during file transfer and installation on controller-1
to fail.

This fix uses post-up configuration to set the correct MTU value to
mimic CentOS's ifup tool behavior.

Test Plan:
    PASS: check that the VLAN's MTU is correct
    PASS: installation on standard lab

Closes-Bug: 1991582

Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>
Change-Id: Id898a0eb132abe6838ddc81ff0adb4401c33d731