Fix cert-mon PriorityQueue regression in python3
In python3 the PriorityQueue raises an exception
due to

TypeError: '<' not supported between
        instances of 'SubcloudAuditData' and 'SubcloudAuditData'

The fix is to include a __lt__ method in SubcloudAuditData.
A timestamp field is added (primarily used in the tuple added to the
queue, but easy enough to include here) in order to aid in the sorting.

Test Plan:

PASS: trigger cert-mon audit for subclouds. Verify that the exception
is not raised, and that subclouds are properly enqueued for audit.

Closes-Bug: 1992680
Change-Id: Ibaa9a421eb809edc434793bc7e8ae92691be021f
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>