Debian: fix VLAN creation in duplex-direct mode
In duplex-direct mode, the ifupdown VLAN configuration tries to
disable DAD using pre-up command 'sysctl -wq net.ipv6.conf.%s.accept_dad=0'
where %s is the vlan device and fails because the VLAN is not created yet.

This happens because Debian's ifupdown tool runs pre-up commands from
/etc/network/interface.d/ before running /etc/network/if-pre-up/vlan
script that creates the VLAN.

Moving the sysctl command to up or post-up won't avoid DAD to run.

The solution found is to add a VLAN creation command in pre-up before
the sysctl command.

Some bonding configuration was also being added to VLANs. Although it
was not causing any issues, it was fixed to be added only to the
bonding interfaces.

Test Plan:
    PASS: Run new VLAN unit test for duplex-direct mode
    PASS: Install duplex-direct system

Closes-Bug: 1992832

Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>
Change-Id: Icbdd113e6051d584959039e85f2dfa078ec0eb2c