Fix Debian issue parsing lldp port power
If  lldpcli -f json0 show neighbor detail
contains ports with power defined for them, the parsing
of that data will fail in lldp sysinv agent.

This is because lldp 1.0.11 seems to omit some of the json
data (pairs and class) if power is not supported/enabled
for that port.

Updating the unit test data to include a port with power data.
Updated the debian test case (number of neigbors and ttl)
based on this new sample data.

Closes-Bug: #1993857
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Icea4bfaaf9b50100296e4268b2b3ca255bd273e5