Increased timeout of agent manager
In some situations the installation is failing due to a timeout waiting
for inventory to complete.

This timeout is being caused because the agent is not having enough time
to communicate with the conductor. In order to correct this the timeout
that agent tries to communicate with the controller was increased from 5
min to 10 min.

Also in case the timeout the agent will keep trying communicate instead
of waiting for the next audit cycle.

TEST PLAN:
PASS: Build, install and verify logs.

Story: 2010087
Task: 45983

Depends-On: https://review.opendev.org/c/starlingx/config/+/852666
Signed-off-by: Fabio Studyny Higa <fabio.studynyhiga@windriver.com>
Change-Id: I5134fbdc6f6644803d69bf99567691f3bf85168c