Fix data-migration on script 98-sysinv-k8s-param-upg
During the upgrade from 22.06 to 22.12, the
k8s control plane component configuration is
saved and restored. Also, the 22.06 configuration
must be loaded into the sysinv database because,
since 22.12, k8s custom configuration is supported
at runtime.

The kubelet component uses a different API that
supports, among others, dictionaries as a type.
The error occurs saving this type of data.

kubelet runtime configuration is not yet supported,
it will be addressed in task:44586 / story:2009766,
so the kubelet section can be safely removed from
the upgrade script at this moment.

Test Plan:
  - Upgrade from centos 22.06 to 22.12 SX and DX

Closes-Bug: 1994884

Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>
Change-Id: Ic560315dc66bc991fe9a98043ef9eeac4fba6441