Fix Controller-1 installation failure due to missing certificate
During pxeboot installation of controller-1, a certificate
dc-adminep-root-ca.crt is not copied to the certs directory as
it does not exist. Controller-1 comes up, but is always locked
and offline, as a result.

This is fixed here by creating the said directory in the
controller_config file.

Test Plan:
PASS: Verify that installation passes once the directory is
      created and the certificate is present in this directory.

Closes-Bug: 1995291

Change-Id: If74e6310f03b795ca632b60c34b78ed7f2d2b349
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>