Create the admin network in sysinv / DB.
The current restrictions for network reconfiguration
post-install is problematic for correcting or updating
network configuration for Subcloud deployments or
supporting Subcloud rehoming procedures.

The proposed solution is to decouple the Distributed
Cloud networking operations from the management network
by introducing a new admin network type that will be
used for the addressing and routing between the central
System Controller and the Subcloud.

Test Plan: Debian
PASS: - system network-add
      - system network-delete
      - system addrpool-add
      - system addrpool-modify
      - system interface-nework-assign
      - system host-addr-list controller-0
      - system interface-network-assign controller-1 (AIO-DX)

Story: 2010319
Task: 46688

Signed-off-by: Mayank Patel <mayank.patel@windriver.com>
Change-Id: I1e7cb782e683daeb847e4855aeb3922a7655b2e9