Improve log messages on sysinv-api ocf script
When sysinv-api fails to initialize, for instance during a swact,
currently it is hard to keep track of the flow of the events as
there will be some instances of the ocf script trying to raise
and check the service status at the same period.

This commit includes some logging on the script to give better
visibility of sysinv-api service status.

Test Plan
PASS: run AIO-DX upgrade successfully
PASS: run AIO-SX upgrade successfully
PASS: install/bootstrap/unlock

Closes-bug: 1995713

Change-Id: Ia93b269b74f84fb4a8ccfe0ee6d1f30f7fdbcff3
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>