Avoid masking err while creating self-signed cert
Modified shell command to avoid succeeding in case of an error in
the creation of self-signed certificate for HTTPS, which leads into
a posterior failure to open the certificate and gives no precise way
for debugging.

Test Plan:
PASS: Enable https for the first time in the system. The command
      should exhibit the system info with https_enabled set to "True",
      plus a message "HTTPS enabled with a self-signed certificate.
      This should be changed to a CA-signed certificate with 'system
      certificate-install'.".

Closes-Bug: 1995633

Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>
Change-Id: Id54e9187b65c710814c6a8abdbc026c412c89865