Adjust partition for compute and storage nodes
Storage nodes don't need to migrate the partition on root disk, as
user partitions are not allowed on the root disk. Let the inventory
report to update the new partition layout.

Also made change 22.12 Debian default partition layout for worker
nodes, see depends-on. This change is to adjust the partition layout
for compute and storage nodes. The layout of controller nodes are
captured at runtime from controller-1 when data migration is invoked.

Test Cases:
   Passed 22.06 to 22.12 upgrade with standard controller, storage and
worker nodes.
   Provision standard system with compute nodes.

Story: 20009303
Task: 46768

Depends-on: https://review.opendev.org/c/starlingx/metal/+/865077

Signed-off-by: Bin Qian <bin.qian@windriver.com>
Change-Id: I9d473db33dc8c544579bdde183ee07ba4392b596