AppFwk: Add FluxCD recovery logic for apply operation
Add some robustenss to the app framework. It is observed that the
framework can reach a bad state where a HelmRelease tries to be
applied but fails because the HelmChart is already in failed state.

When applying a HelmRelease FluxCD doesn't trigger a HelmChart
reconciliation, thus HelmChart is in failed state and HelmRelease is
stuck waiting for it.
Framework times-out after 1 hour of waiting for FluxCD to finish the
apply operation, which never happens.

Implement logic to trigger a HelmChart reconciliation.
Implement optimization to trigger HelmRelease reconciliaton right
after HelmChart reconciliation, to skip the time FluxCD waits until
periodic reconciliation is triggered.

Skip CentOS testing because we use the same fluxcd and kubernetes.
The only difference is the python kubernetes library, but the
implementation does not use any new API calls.

Tests on AIO-SX Debian:
PASS: AIO-SX unlocked enabled available
PASS: Can apply app when apply started with a HelmChart
      for a platform-integ-apps in Failed state
PASS: Can apply app when apply started with both a HelmChart and a
      HelmRelease for a platform-integ-apps in Failed state

Closes-Bug: 1996747
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Iea19c5458bb1c6d739dfe0ef94af09c85e90cd9a