Alarm 500.200 "expiring soon" not updating after change
This issue was been caused because when renewing the certificate
it wasn't updating fields, like "Reason Text". To fix it,
this change will delete the expiring soon alarm and replace it by
with the info of the new certificate, if it is necessary raising a new
expiring soon alarm.

Test Plan:

PASS: Renew a certificate with "expiring soon" alarm, and verify if the
      time registered in 'cert-alarm.log' and "Reason Text" in fm-alarm
      list match with the time "Not After" in the certificate
      information.
PASS: Renew a expired certificate with a certificate with less than 3
      days to expire and verify if the expired alarm is deleted
      and replaced by the "expiring soon" alarm.
PASS: Delete the certificate of the "expiring soon" alarm and check if
      the alarm is deleted.

Closes-Bug: 1997037
Change-Id: I0f724566de10e66ab95f6ba862836f1fc4e49f32
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>