Add PSAC labels to only managed namespaces
Pod Security Admission Controller (PSAC)
labels are automatically added to namespaces that applications are
applied in. Sysinv used to assume that all namespaces that
applications are deployed in are managed by sysinv, and assigns
a privileged label to it. This can result in privileged labels
being assigned to the default namespace, for example, if an
application is deployed there. This commit limits PSAC label
assignment to only a list of namespaces that we know are for our
applications.

PASS: Bootstrap system, check that cert manager namespace has
      PSAC labels (label still exist for application deployment)
PASS: Apply test application to default namespace, and ensure
      PSAC labels are not assigned

Closes-Bug: 1997117
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>
Change-Id: I40da7d9d3b2e79d2133390ea7b270e046200b57e